+++ lib/pages/main_page/statistics_widget_option.dart
@@ -1,4 +1,10 @@
+import 'package:fl_chart/fl_chart.dart';
 import 'package:flutter/material.dart';
+import 'package:game_lists/model/game.dart';
+import 'package:game_lists/model/status.dart';
+import 'package:game_lists/pages/main_page/charts/games_count_by_game_status_widget.dart';
+import 'package:game_lists/pages/main_page/charts/games_count_by_year.dart';
+import 'package:hive/hive.dart';
 
 class StatisticsWidgetOption extends StatefulWidget {
   StatisticsWidgetOption({Key? key, this.tabController}) : super(key: key);
@@ -15,7 +21,13 @@ class _StatisticsWidgetOptionState extends State<StatisticsWidgetOption> {
     return TabBarView(
       controller: widget.tabController,
       children: [
-        Icon(Icons.directions_car),
+        ListView(
+          children: [
+            GamesCountByGameStatusWidget(),
+            CompletedGamesCountByFinishYear(),
+            CompletedGamesCountByReleaseYear(),
+          ],
+        ),
         Icon(Icons.directions_transit),
       ],
     );